+++ docs/src/pages/components/loader/examples/index.ts
@@ -1,0 +1,7 @@
+export { DefaultLoaderExample } from './DefaultLoaderExample';
+export { DeterminateLoaderExample } from './DeterminateLoaderExample';
+export { LoaderAccessibilityExample } from './LoaderAccessibilityExample';
+export { LoaderColorExample } from './LoaderColorExample';
+export { LoaderIsPercentageTextHiddenExample } from './LoaderIsPercentageTextHiddenExample';
+export { LoaderSizesExample } from './LoaderSizesExample';
+export { LoaderVariationExample } from './LoaderVariationExample';